pyproject.toml
[tool.poetry]
name = "backend-onboard-youngbin"
version = "0.1.0"
description = ""
authors = ["youngbin.seo <youngbin.seo@a-bly.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.1"
djangorestframework = "^3.11.0"
drf-yasg = "^1.21.7"
pymysql = "^1.1.0"
djangorestframework-simplejwt = "^5.3.0"
redis = "^5.0.0"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
